template_escaping: escaping in templates is modified
the '\' is parsed by lexer and the template_compile
used this character to escaping. It was confusing.
Now, it has to be used $$ for printing literal $ in the result

The other modification is the escaping of the '@' if it follows macro or
temlate function.
if non-numeric value follows the '@', use the '@' as a literal '@', but
show a warning message.
It has to be used '@@' for printing literal '@' in the result

Signed-off-by: Juhasz Viktor <jviktor@balabit.hu>