affile: regular and irregular files handling fix
On windows we deal only with regular files.
On unix platforms we support pipes and other kind of irregular
files[eg.:FreeBSD /dev/klog].
The way we determine if a file is regular or not, that we use stat/fstat
for every message. In order to eliminate stat/fstat overhead we call only
once stat/fstat and store the result.

References: v1 B-09743