templates: use the global options if NULL is passed as template_options
Use the global template-options structure in GlobalConfig instead of a
hard-wired set of defaults when the template_options argument is NULL.

This is a minor change in user visible behaviour, as the global
ts-format(), frac-digits() and time-zone related options will
become effective to the call-sites which didn't provide their own
idea about template options before.

Previously, a hard-wired set of defaults were used, e.g.:
  - frac-digits(0)
  - ts-format(bsd)
  - neither local-time-zone nor send-time-zone are set

Out of these, only frac-digits() have some chance to actually break
user-configurations as the call-sites that used the hard-wired values
are not really likely to use time related macros.

The confusion of not applying the global options is potentially larger,
not to mention the code complexities this causes in all drivers using
templates.

These are the call-sites that used the hard-wired values before:
  - filter expressions that use the comparison operators
  - regexp based substitutions
  - incoming templates for parsers
  - the new value rewrite rules
  - smtp values
  - the just introduced drivers (stomp, amqp)
  - patterndb context-id

The consequence of this patch that we're going to have a LogTemplateOptions
available in template functions too.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>