logreader-logwriter: fix syslog-ng hang during reload on Windows
During reload in the main_loop_io_worker_sync_call the
main_loop_io_workers_quit flag is set which can prevent the processing
of the socket events in logreader and logwriter.

This can happen the following way:
In iv_fd_got_event function in compat.c the Windows' internal network
event records are cleared as the event object become non-signaled
because of the call to WSAEnumNetworkEvents function so the network
event is assumed to be processed.
The logreader and logwriter starts to processing the network event
initiated by the iv_fd_got_event. If the main_loop_io_workers_quit
flag is set during this operation (as it set handling the reload)
the processing can be interrupped.

In bi-directional communication cases (for example the RLTP protocol)
leaving the socket event unprocessed can lead to a dead-lock situation,
since both communication sides will be waiting for the other side to
answer:
one has already sent the message but the other does not process it
since the reload interrupts the processing. The receiver side would go
on only if it had received an other network event but this will not
happen because the sender is waiting for reply too. This dead-lock
situation will finish only when the sender's timeout exceeds and it
decides to terminate the connection. In this case it sends a FIN TCP
messages to the receiver which wakes up suddenly. It sends the awaited
reply but it is already late because the other side has started to
tear down the TCP connection.

To fix this problem a flag is set in init functions of both logreader
and logwriter (that is called after reload) on Windows and if the flag
is set we will try to process network messages (triggered by
update_watches) regardless of weather there are any messages to be
processed or not.

Signed-off-by: Daniel Gados <gdani@balabit.hu>