log_queue_fifo: log_queue_fifo_is_empty method is implemented
It was required, because there was possible to loss and leak message
in case of using logqueue-fifo. The symptom is the following:

- The reader fill its own input queue in its fetch loop,
while the fetch loop is running its input queue won't be synchronized out.
While it isn't synchronized, the get_length won't count messages in the this input queue.
But if the reap timer elapsed and the driver ask the writer,
whether it has pending writes, the writer will answer, that it hasn't pending writes,
so the queue will be released and the messages in the input queue will be loss and leaked.

This solution still has race, but much better than the original state.
The good solution should be, that increment the stored message counter in
the push_tail side, not in the synchronize side.
It causes about 10-20% performance reducing.

Signed-off-by: Juhasz Viktor <jviktor@balabit.hu>