tlstransport: fixing hang on Windows
In some cases we do not recognize that the TLS/SSL connnection was
dropped on Windows that causes hang in the RLTP client (the server
is restarted but the client tries to write whitout noticing the
problem). This happens because when the SSL_read and SSL_write
functions return with error and the SSL error is SSL_ERROR_SYSCALL
the errno is set only on Unix like systems, while on Windows errno
still stores the previous value. The soulution is to explicitly
set errno on Windows to the value returned by WSAGetLastError.

Signed-off-by: Daniel Gados <gdani@balabit.hu>