logreader: fix incorrect SDATA part of syslog messages (missing file position)
It happened becaues the logreader's size memeber was zero while reading
a file on Windows. An is_regular member was introduced to distinguish
between file sources and other sources. From now if the size memeber
is zero for file sources while reading the file we will check the size
again. The is_file_regular function was moved from logwriter.c to misc.c.

Signed-off-by: Daniel Gados <gdani@balabit.hu>