basic-proto: substitute the invalid bytes according to the encoding with a question mark
Previously the syslog-ng hung when the encoding was set and an invalid
byte was read according to the set encoding.
From now instead of the invalid bytes a question mark will appear at the
output.
Technically syslog-ng is skipping the same number of bytes in the buffer as
the question mark is represented in the original encoding to ensure that
after the reverse conversion the length of the text will be the same as
the original text. (This is requied for file sources when splitting occurs
and the eol is not read yet.)

Signed-off-by: Daniel Gados <gdani@balabit.hu>