modules: Fix the initialisation of LogThrDestDriver modules
All of the LogThrDestDriver modules were calling
log_threaded_dest_driver_init_method() first thing, but that depends on
variables set later, leading to crashes on reload. This patch moves the
call of said method to the end of the init functions, just before the
thread start call. This also requires the _init_method() to be split: we
need to call the LogDestDriver init method at the start of the module
_init(), so that was taken out of LogThrDestDriver, and moved to the
destinations.

Reported-by: Balazs Scheidler <bazsi@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>