test_dnscache: remove inet_ntop() benchmark
This is not needed anymore, it was used to determine whether it makes
sense to store negative matches in the DNS cache. As inet_ntop()
proved about 5x slower it did make sense. But we are doing that and
no need to validate this decision at every unit test invocation. I don't
expect this to change dramatically.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>