tags: update stats counters at reload time
When a tag is registered, the final value for stats-level() may not
be known. Also, when the stats-level() is decreased with a reload,
already registered counters might need to be unregistered.

This is solved by a log_tags_reinit_stats() function call, that is invoked
at cfg_init() time, this then in turn iterates on all registered
tags and registers/unregisters counters as needed.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>