host-resolve: empty string in resolve_hostname_to_sockaddr() resolves to the wildcard address
As call-sites had to handle the default bind address on their own,
a better strategy is to handle a specific hostname to resolve to the
wildcard address that can be used to bind a service.

With IPv4, using the empty string or "0" literal works. With IPv6
unfortunately neither does. This patch adds this capability
to resolve_hostname_to_sockaddr(), either the empty string or a NULL
value returns the 0.0.0.0:0 or [::]:0 for IPv4 and IPv6 respectively.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>