host-resolve: add resolve_hostname_to_hostname() function
This function can be used to reformat a hostname according to use-fqdn() &
normalize-hostnames() settings. Basically if we got a hostname without IP
address information, this is the best bet we can do to format a hostname
that best matches the one we would be generating if we created a message.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>