LogProtoServer: fixed a use-after-free after reload
self->options is a borrowed memory area, managed by the owning driver
that constructed the given LogProtoServer instance. During reload however,
the driver object itself is freed, while the LogProtoServer instance
is kept alive, thereby causing the options pointer to point into the
void.

The solution is to update the options pointer whenever the owner is changed,
chaining into a similar solution in the LogReader code which does exactly
that.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>