cfg-grammar: create CHECK_ERROR() variant without error message
In some cases CHECK_ERROR() macro is invoked without an error message,
which simply means it should simply return the grammar without producing
a verbose error message. This happens when a recursive grammar already
produced a meaningful error message.

Instead of overloading CHECK_ERROR() with a NULL errormsg in this case,
introduce a separate macro.

This also fixes a warning when compiling with clang.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>