serialize: use gssize for the result of I/O operations
fread/fwrite returns signed quantity and anything below 0 indicates
a failure. This code errorenously used gsize instead of gssize causing
the error handling never to be triggered.

Fortunately this code is only being used to upgrade a persist-state file
in the older format, and read() rarely if ever fails once the file could
successfully be opened.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>