mainloop: factor out systemd communication into a service-management module
Some communication takes place between syslog-ng and systemd after
the former has started. We indicate when initialization at startup
finished and also some textual messages to indicate our state.

Earlier it was right in the middle of the mainloop code, now factor it
out to a separate module, perhaps this interface can be useful
for other similar implementations.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>