apphook: add per-thread start/stop functions
This patch adds two functions to apphook, one to be invoked at every
thread creation and its pair to be invoked when a thread is about to exit.

This decreases the level of coupling on code that actually spawns threads.
A dynamic registration would be even nicer, but this is what we have for
now.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>