mainloop-worker: remove worker thread signal setup
We used to require SIG_BLOCKing a number of signals in worker threads as
we pretty much relied on SIGHUP/SIGTERM/SIGCHLD/SIGINT to arrive to the
main thread. This is not guaranteed by POSIX however.

In POSIX signals are delivered to a process, and the OS is free to choose
a random thread to deliver that signal to, assuming the signal is not
blocked there.

That was the reason we blocked these signals everywhere except the
main thread.

Stuff however changed when we switched to a multithreaded architecture and
started using libivykis. ivykis pretty much handles this just fine by
rerouting signals to callbacks in the thread that registered the interest.

That way, it doesn't really matter which thread the signal is delivered
to it is taken care of.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>