autogen.sh: Allow extra code in a separate, optional file
When Mk/autogen-extras.sh exists, source it, so it can do some extra
setup. The main reason for this split is that there are setup steps that
we do not need in OSE, and keeping a patch against autogen.sh would
easily lead to conflicts. Putting it into a separate file makes
maintenance easier.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>