stats: "source" member & variables renamed to component
This patch makes it easier to understand what source refers to, as the
"source" name is overloaded a number of times. Basically, counters
are related to the component that registered them and this relation is
kept around and published where statistics are published.

This patch also moves the functions related to formatting component names
and counter types to stats-cluster where they belong.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>