value-pairs: Ignore empty values in SDATA too
The value-pairs framework ignored empty values when they were explicitly
specified as such, but when they came through via SDATA or any other
means, it treated them as empty strings, and allowed them through. This
was inconsistent, and a waste too.

This patch corrects that behaviour, and empty-valued SDATA members will
not pass through anymore.

Reported-by: Lucas McLane <lucas@clicksecurity.com>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>