confgen: backticks weren't processed in generated configuration part
produced by confgen module. The reason of this problem was,
that the cfg_lexer_subst_args was called only for read lines
from configuration files (YY_INPUT) and for the processed blocks.

If process the backticks for a cfg input buffer is a must,
the cfg_lexer_subst_args has to be called explicit.
Now the confgen module call it before add the buffer to the lexer

Signed-off-by: Juhász Viktor <jviktor@balabit.hu>