value-pairs: Restructure vp_walker_name_split()
Instead of splitting tokens into an array of gchar*s, use a GPtrArray
instead, so that vp_walker_name_value_split() does not have to copy from
a pointer array to a string array.

Also introduce vp_walker_name_combine_prefix(), a short function to
replace the g_strjoinv() we used before.

The advantage of this is that it's much easier to extend later, for
example with array support. The code of vp_walker_name_split() becomes
easier to follow too.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>