afmongodb: Add/restore authentication support
A long, long time ago, the MongoDB destination was able to authenticate
when connecting to MongoDB. This was broken soon after, and disabled,
with the username() and password() options going away in February this
year.

This patch brings these options back, and implements the authentication
functionality too.

While most of it will work with a libmongo-client older than 0.1.7, if
the driver falls back to a replica, it will not work until restarted.
When using libmongo-client 0.1.7+, that works too. For this reason, the
internal copy of libmongo-client is also updated for the fresh 0.1.7
release.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>