transport: introduce LogTransportAuxData struct
The purpose of this struct to encapsulate transport specific meta
information that relates to a given read operation. Some transports
(like unix-domain sockets) provide additional information that is useful
for syslog-ng, like the PID of the sending process, etc.

This struct makes it possible to carry this kind of meta information
in a format that allows injecting these values as name-value pairs to
messages. The peer GSockAddr address also got its new home in the struct.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>