cfg-tree: Added a few tests for cfg-tree
These are some simplistic tests for cfg-tree, that verify some of the
functionality provided. In particular, this tests cfg_tree_start() and
cfg_tree_stop() in a few conditions, using only a simplistic LogPipe.

The intent is to check that cfg_tree_start() initializes only those
pipes that can be and nothing after, and cfg_tree_stop() deinits those
that were initialized before.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>