mainloop: Deinit the new config when reverting to the old one
When reloading, and the new config fails, deinit it before initializing
the old config. This is so that conflicting things do not remain held by
the half-initialized new config, while the old tries to take it
over. (It also removed a couple of memory leaks, most likely.)

The reason we can do this, is because cfg_tree_stop() (called by
cfg_deinit()) goes over all the known nodes, and log_pipe_deinit() is
also smart enough to not deinit a node that has not been inited before.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>