cfg: add use-rcptid() global option
This patch adds a use-rcptid() global option that causes syslog-ng to assign
a reception ID to be assigned to all messages received and generated by syslog-ng.

This can be accessed using the $RCPTID macro and is unique on a given host.
The counter wraps around after 48 bits and is never zero.

It can be used to generate a unique ID for a message.

Signed-off-by: Csaba Karsai <cskarsai@balabit.hu>
Signed-off-by: Viktor Juhasz <jviktor@balabit.hu>
Signed-off-by: Viktor Tusa <tusa@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>