rcptid: Add unique receipient id framework
This patch adds an rcptid subsytem that can be used to generate per-message
identifiers unique on a given host.

This id is persistent across restarts, and is globally unique for the messages
generated/received on this host/instance.  It is somewhat faster than uuid
template function, but is only a non-zero 48 bit counter.

Signed-off-by: Csaba Karsai <cskarsai@balabit.hu>
Signed-off-by: Viktor Juhasz <jviktor@balabit.hu>
Signed-off-by: Viktor Tusa <tusa@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>