pdbtool: If a test message has no program attribute, warn when verbose
When encountering a test message that has no program attribute, warn
about it when verbose mode is on. The rationale for only doing it in
verbose mode is that there are valid reasons for having no program
attribute for a test (to temporarily disable it, for example), and we
don't want to needlessly spew a lot of warnings at the user.

Reported-by: Fabien Wernli <bugzilla.balabit@faxm0dem.org>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>