LogProto/LogTransport: use LogTransportAuxData to pass meta information around
Earlier, syslog-ng didn't have the notion of transport/logproto specific
meta information. The only thing that was passed around when receiving
a message is the address of the peer as a GSockAddr instance.

This patch is changing this, it uses the introduced LogTransportAuxData
struct, which encompasses all information that the LogTransport/LogProto
combination is willing to provide.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>