doc/Makefile.am: Always include the manpages in the tarball
Whether man page building is enabled or not, include them in the
tarball. This means that you won't be able to build a tarball from git
unless you build the man pages too, but it also means that if you do a
make dist from a tarball, without xsltproc installed, they will still be
included.

Reported-by: Peter Czanik <czanik@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>