Makefile.am: bison >= 3 improvements
We add -Wno-yacc to AM_YFLAGS, to simplify the Makefile. This works for
all autoconf versions, with all bison versions, and we require bison
anyway. We also add -Wno-other there, to skip the bison warnings we
previously grepped out in Mk/lex-rules.am.

Grepping does not work anymore, as bison 3.0 uses multi-line messages,
so we get half of them anyway.

This addresses another part of #49.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>