cfg-grammar.y: remove the use of YYID
Bison3 dropped the YYID macro, which was used by an overridden
YYLLOC_DEFAULT implementation that we use.

@algernon has tried to fix this in eeee0a22d, however while it worked
with bison3 it caused warnings with older bison versions.

Looking into the matter, it seems that YYID() is only needed to suppress
lint warnings, which we don't use in our toolchain and even bison has
ceased supporting it, so the better solution is to remove YYID references
from our code completely.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>