cfg-lexer-subst: report an error if backtick expansion happens after backslash in a string
If we are substituting a string within double-quoted string and the
last character was a quote character, we can't really decide what to do.

The backslash character could protect the backtick, but double-backticks should
be used for those.

If we simply substituted the backticked variable, the backslash would apply
to its first character, which is almost certainly not what the user wanted.

Report an error in this case instead.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>