cfg-lexer: don't take ownership of the buffer argument in cfg_lexer_include_buffer()
cfg_lexer_include_buffer() assumed ownership of the input buffer, e.g.
it took care of freeing it. This was a bit unintuitive and cumbersome to
use and this code wasn't performance critical anyway.

This patch changes the function to only get a borrowed reference and the
responsibility of freeing the buffer stays at the caller.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>