hostname: fix abort when the local hostname is unknown to both DNS and /etc/hosts
When the local hostname is not an FQDN, syslog-ng consults the local
resolver to get a fully qualified domain name from either the DNS or the
hosts file.

Previously if this failed, syslog-ng aborted after printing an error
message. This patch changes behaviour so that in this case syslog-ng
will happily work with the non-qualified hostname after printing
a notice to the internal() log.

This was triggered by a SUSE test environment, so this actually happens.

Reported-By: Peter Czanik <czanik@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>