hostname-unix: find the first non-FQDN name, instead of always picking the primary name
The primary name of a host should always be FQDN, especially when it is
resolved from DNS. Even the hosts file should list the FQDN first. With
this assumption, syslog-ng used to pick the primary name, always.

So with an inconsistent configuration, where /etc/hosts listed the FQDN
at the non-first position, syslog-ng picked the wrong name.

This patch changes the logic and finds the first FQDN. This patch also
refactors hostname-unix a bit and covers the hostname extraction function
with testcases.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>