control: Fix the test_control_cmds test to be deterministic
The order in which stats are generated is not deterministic, therefore
testing whether the generated stats are the same as a pre-determined
list will sooner or later fail. There are two things we can test: that
the first line is the stats header, and that the number of returned
lines are of the expected number. In this case, I believe it is enough
if we test the header only.

Reported-by: Andras Mitzki <micek@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>