json: add extract-prefix() option to json-parser
This option adds a new option to json-parser() that instructs the parser
to only extract JSON members from a specific subtree of the incoming object.

Example:
  json-parser(extract-prefix("foo.bar[5]"));

Assuming that the incoming object is named msg, this is equivalent to the
following javascript code:

  msg.foo.bar[5]

The resulting expression must be a JSON object, so that syslog-ng can
extract its members into LogMessage name-value pairs.

This also works when the top-level object is an array, as extract-prefix()
allows the use of an array index at the first indirection level, e.g.

  json-parser(extract-prefix("[5]"));

which translates to: msg[5]

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>