contrib: Add a Valgrind suppression file
When debugging memory leaks in syslog-ng, there are a couple of known
false positives, or leaks we do not care about (such as OpenSSL
initialisers and such). To silence those, a suppression file was
created. Simply add a --suppressions=/path/to/valgrind/syslog-ng.supp to
the Valgrind command line, and there will be so much less noise.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>