Makefile: Fix make dist in certain edge cases
Due to the way internal modules were built, if both mongodb and
libmongo-client was disabled, make dist failed. When amqp was explicitly
disabled, make dist failed.

The reason being, we use a dummy.c to trigger building the bundled
libraries, but the dummy.c files were not created if mongodb/amqp were
disabled. Now we create them anyway, just don't recurse into the
directory when the module is disabled.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>