stats: Fixed center statistics
The "center" statistics is a cumulative counter for whole syslog-ng.
Unfortunately, this feature was left log time ago.

This fix based on OSE 3.6, but some case are missing there. The "center"
statistics was implemented in ...queue_method in driver.c originally, but
sometimes this was overlapped in source and destination drivers.

FYI:
 * center/processed means count of all received messages by source drivers
 * center/queued means count of all queued messages by destination drivers

References: v1 D-08043
References: kayako 2008
References: bugzilla #31194