mainloop: Deinit the new config when reverting to the old one
When reloading, and the new config fails, deinit it before initializing
the old config. This is so that conflicting things do not remain held by
the half-initialized new config, while the old tries to take it
over. (It also removed a couple of memory leaks, most likely.)

The reason we can do this, is because cfg_deinit() goes over only also
inited pipes.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>
Signed-off-by: Juhász Viktor <jviktor@balabit.hu>