afsocket: avoid persist_name usage duplication in the same configuration.
persist name of the network based drivers is built from the server address
and port. If there are two network based destination with same port
and server address in a configuration, it means, that their persist name
is the same. It can cause problem for example in case of disk queue usage.

Whit this patch syslog-ng detects this collision and stop it before bigger
errors are occured.

Signed-off-by: Juhász Viktor <jviktor@balabit.hu>

References: bugzilla #31363