cfg_reinit: implementing of reinit interface to GlobalConfig and LogCenter.
In case of reload, when a new configuration initialization is failed,
The cfg_init called for the old configuration. But the cfg_init called
the log_center_init and log_center_init called log_center_init_pipe_line
But it mustn't becuase this calling duplicates the initialized pipes and
the children of multiplexers.
This patch use cfg_reinit and log_center_reinit and in this case
the pipe line initializations are skipped, avoid duplications.

Signed-off-by: Juhász Viktor <jviktor@balabit.hu>

Conflicts:
	lib/cfg.c
	lib/mainloop.c

References: bugzilla #31447