afinter: Added internal queue length statistics
It can really help cases, when internal() driver is not used, and internal
queue starts to queue up. The best would be not posting when no internal()
is present, and this change whould be rather trivial, but I thought there
is a reason why we work like this, so I didn't change that.

Signed-off-by: Tusa Viktor <tusavik@gmail.com>