logsource: moving pos_tracking flag from LogSourceOptions to LogSource to avoid use after free
After configuration reload the source driver gets freed but the LogSource
(LogReader) is not because until the last message gets acked the message
holds a reference to the LogSource.
When the source driver is freed during reload the contained LogSourceOptions
(or LogReaderOptions) structure is freed as well but the LogSource holds a
pointer to the source driver's LogSourceOptions thus after the source driver
is freed the LogSource's options points to freed and possibly messed memory.
Then later when the message gets acked the log_source_msg_ack function is
called which checks the already freed and possibly messed LogSourceOptions'
flags. If the position tracking originally was not set in this bitfield but
after the memory is messed up the flag happens to be set then the
log_source_msg_ack tries to write in the LogSource's ack_list leading to
null pointer dereference.
After the modification the flags is removed from the LogSourceOptions and
it is moved as pos_tracking boolean member to the LogSource avoiding both
use after free and null pointer dereference.

Signed-off-by: Daniel Gados <gdani@balabit.hu>