plugin.c: Use G_MODULE_BIND_LAZY instead of G_MODULE_BIND_LOCAL
When we have a situation where we load a syslog-ng module, linked
against a certain library, and that module tries to load a plugin that
requires the library the module is linked against, but the plugin isn't,
using BIND_LOCAL will result in symbol lookup errors. Not using it will
correctly resolve the symbols previously loaded.

This is needed to be able to load Lua, Perl, etc modules.

On some exotic platforms (like AIX), where the linker tends to act
strangely, this may cause problems. If it does, we should likely use
BIND_LOCAL there.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>