logproto: Added suffix based multiline mode.
Suffix-based multiline is very similar to prefix-garbage based, except it
appends the garbage part to the message as well. With this way, the user can
workaround the absence of multi-line timeout feature.

The differences: The old multi-line-mode(regexp) is still usable, but there is a new alias
for it, multi-line-mode(prefix-garbage). And the entirely new mode is
multi-line-mode(prefix-suffix). The user can use multi-line-suffix() instead of multi-line-garbage().

Requested-by: Evan Rempel <erempel@uvic.ca>
Signed-off-by: Tusa Viktor <tusavik@gmail.com>