tlscontext: disable TLSv1.2 protocol in case of cryptoAPI engine is used.
It is required, because in case of 1.2 protocol and mutual authentication
the client tries to sign the hash its private key.
But the algorithm of the private key can be very various and not the all
can be handled in the capi engine, e.g.: sha512 doesn't work well.
In case of v1.1 and v1.0 the sign is made by sha1_md5.

Signed-off-by: Juhász Viktor <jviktor@balabit.hu>

References: bugzilla #31538